+++ examples/image_example.py
@@ -1,3 +1,5 @@
+import tkinter
+
 import customtkinter
 import os
 from PIL import Image