+++ customtkinter/windows/ctk_tk.py
@@ -1,10 +1,10 @@
 import tkinter
-from distutils.version import StrictVersion as Version
 import sys
 import os
 import platform
 import ctypes
 from typing import Union, Tuple, Optional
+from packaging import version
 
 from .widgets.theme import ThemeManager
 from .widgets.scaling import CTkScalingBaseClass
@@ -12,8 +12,10 @@ from .widgets.appearance_mode import CTkAppearanceModeBaseClass
 
 from customtkinter.windows.widgets.utility.utility_functions import pop_from_dict_by_set, check_kwargs_empty
 
+CTK_PARENT_CLASS = tkinter.Tk
 
-class CTk(tkinter.Tk, CTkAppearanceModeBaseClass, CTkScalingBaseClass):
+
+class CTk(CTK_PARENT_CLASS, CTkAppearanceModeBaseClass, CTkScalingBaseClass):
     """
     Main app window with dark titlebar on Windows and macOS.
     For detailed information check out the documentation.
@@ -35,7 +37,7 @@ class CTk(tkinter.Tk, CTkAppearanceModeBaseClass, CTkScalingBaseClass):
         self._enable_macos_dark_title_bar()
 
         # call init methods of super classes
-        tkinter.Tk.__init__(self, **pop_from_dict_by_set(kwargs, self._valid_tk_constructor_arguments))
+        CTK_PARENT_CLASS.__init__(self, **pop_from_dict_by_set(kwargs, self._valid_tk_constructor_arguments))
         CTkAppearanceModeBaseClass.__init__(self)
         CTkScalingBaseClass.__init__(self, scaling_type="window")
         check_kwargs_empty(kwargs, raise_error=True)
@@ -241,16 +243,16 @@ class CTk(tkinter.Tk, CTkAppearanceModeBaseClass, CTkScalingBaseClass):
     @classmethod
     def _enable_macos_dark_title_bar(cls):
         if sys.platform == "darwin" and not cls._deactivate_macos_window_header_manipulation:  # macOS
-            if Version(platform.python_version()) < Version("3.10"):
-                if Version(tkinter.Tcl().call("info", "patchlevel")) >= Version("8.6.9"):  # Tcl/Tk >= 8.6.9
+            if version.parse(platform.python_version()) < version.parse("3.10"):
+                if version.parse(tkinter.Tcl().call("info", "patchlevel")) >= version.parse("8.6.9"):  # Tcl/Tk >= 8.6.9
                     os.system("defaults write -g NSRequiresAquaSystemAppearance -bool No")
                     # This command allows dark-mode for all programs
 
     @classmethod
     def _disable_macos_dark_title_bar(cls):
         if sys.platform == "darwin" and not cls._deactivate_macos_window_header_manipulation:  # macOS
-            if Version(platform.python_version()) < Version("3.10"):
-                if Version(tkinter.Tcl().call("info", "patchlevel")) >= Version("8.6.9"):  # Tcl/Tk >= 8.6.9
+            if version.parse(platform.python_version()) < version.parse("3.10"):
+                if version.parse(tkinter.Tcl().call("info", "patchlevel")) >= version.parse("8.6.9"):  # Tcl/Tk >= 8.6.9
                     os.system("defaults delete -g NSRequiresAquaSystemAppearance")
                     # This command reverts the dark-mode setting for all programs.