+++ customtkinter/windows/widgets/ctk_scrollable_frame.py
@@ -1,4 +1,4 @@
-from typing import Union, Tuple, Optional
+from typing import Union, Tuple, Optional, Any
 try:
     from typing import Literal
 except ImportError:
@@ -18,7 +18,7 @@ from .theme import ThemeManager
 
 class CTkScrollableFrame(tkinter.Frame, CTkAppearanceModeBaseClass, CTkScalingBaseClass):
     def __init__(self,
-                 master: any,
+                 master: Any,
                  width: int = 200,
                  height: int = 200,
                  corner_radius: Optional[Union[int, str]] = None,
@@ -43,6 +43,7 @@ class CTkScrollableFrame(tkinter.Frame, CTkAppearanceModeBaseClass, CTkScalingBa
         # dimensions independent of scaling
         self._desired_width = width  # _desired_width and _desired_height, represent desired size set by width and height
         self._desired_height = height
+        self._border_width = border_width if border_width else 0
 
         self._parent_frame = CTkFrame(master=master, width=0, height=0, corner_radius=corner_radius,
                                       border_width=border_width, bg_color=bg_color, fg_color=fg_color, border_color=border_color)
@@ -99,10 +100,11 @@ class CTkScrollableFrame(tkinter.Frame, CTkAppearanceModeBaseClass, CTkScalingBa
         border_spacing = self._apply_widget_scaling(self._parent_frame.cget("corner_radius") + self._parent_frame.cget("border_width"))
 
         if self._orientation == "horizontal":
+            border_padding = (0, self._border_width +1)
             self._parent_frame.grid_columnconfigure(0, weight=1)
             self._parent_frame.grid_rowconfigure(1, weight=1)
             self._parent_canvas.grid(row=1, column=0, sticky="nsew", padx=border_spacing, pady=(border_spacing, 0))
-            self._scrollbar.grid(row=2, column=0, sticky="nsew", padx=border_spacing)
+            self._scrollbar.grid(row=2, column=0, sticky="nsew", padx=border_spacing, pady=border_padding)
 
             if self._label_text is not None and self._label_text != "":
                 self._label.grid(row=0, column=0, sticky="ew", padx=border_spacing, pady=border_spacing)
@@ -110,10 +112,11 @@ class CTkScrollableFrame(tkinter.Frame, CTkAppearanceModeBaseClass, CTkScalingBa
                 self._label.grid_forget()
 
         elif self._orientation == "vertical":
+            border_padding = (0, self._border_width +1)
             self._parent_frame.grid_columnconfigure(0, weight=1)
             self._parent_frame.grid_rowconfigure(1, weight=1)
             self._parent_canvas.grid(row=1, column=0, sticky="nsew", padx=(border_spacing, 0), pady=border_spacing)
-            self._scrollbar.grid(row=1, column=1, sticky="nsew", pady=border_spacing)
+            self._scrollbar.grid(row=1, column=1, sticky="nsew", padx=border_padding, pady=border_spacing)
 
             if self._label_text is not None and self._label_text != "":
                 self._label.grid(row=0, column=0, columnspan=2, sticky="ew", padx=border_spacing, pady=border_spacing)
@@ -160,7 +163,8 @@ class CTkScrollableFrame(tkinter.Frame, CTkAppearanceModeBaseClass, CTkScalingBa
             self._create_grid()
 
         if "border_width" in kwargs:
-            self._parent_frame.configure(border_width=kwargs.pop("border_width"))
+            self._border_width = kwargs.pop("border_width")
+            self._parent_frame.configure(border_width=self._border_width)
             self._create_grid()
 
         if "fg_color" in kwargs:
@@ -181,10 +185,10 @@ class CTkScrollableFrame(tkinter.Frame, CTkAppearanceModeBaseClass, CTkScalingBa
             self._scrollbar.configure(fg_color=kwargs.pop("scrollbar_fg_color"))
 
         if "scrollbar_button_color" in kwargs:
-            self._scrollbar.configure(fg_color=kwargs.pop("scrollbar_button_color"))
+            self._scrollbar.configure(button_color=kwargs.pop("scrollbar_button_color"))
 
         if "scrollbar_button_hover_color" in kwargs:
-            self._scrollbar.configure(fg_color=kwargs.pop("scrollbar_button_hover_color"))
+            self._scrollbar.configure(button_hover_color=kwargs.pop("scrollbar_button_hover_color"))
 
         if "label_text" in kwargs:
             self._label_text = kwargs.pop("label_text")