+++ customtkinter/windows/widgets/theme/__init__.py
@@ -4,6 +4,6 @@ from .theme_manager import ThemeManager
 try:
     ThemeManager.load_theme("blue")
 except FileNotFoundError as err:
-    raise FileNotFoundError(f"{err}\n\nThe .json theme file for CustomTkinter could not be found.\n" +
+    raise FileNotFoundError(f"{err}\nThe .json theme file for CustomTkinter could not be found.\n" +
                             f"If packaging with pyinstaller was used, have a look at the wiki:\n" +
                             f"https://github.com/TomSchimansky/CustomTkinter/wiki/Packaging#windows-pyinstaller-auto-py-to-exe")