+++ customtkinter/windows/widgets/theme/theme_manager.py
@@ -1,5 +1,6 @@
 import sys
 import os
+import pathlib
 import json
 from typing import List, Union
 
@@ -15,7 +16,8 @@ class ThemeManager:
         script_directory = os.path.dirname(os.path.abspath(__file__))
 
         if theme_name_or_path in cls._built_in_themes:
-            with open(os.path.join(script_directory, "../../../assets", "themes", f"{theme_name_or_path}.json"), "r") as f:
+            customtkinter_path = pathlib.Path(script_directory).parent.parent.parent
+            with open(os.path.join(customtkinter_path, "assets", "themes", f"{theme_name_or_path}.json"), "r") as f:
                 cls.theme = json.load(f)
         else:
             with open(theme_name_or_path, "r") as f:
@@ -34,6 +36,12 @@ class ThemeManager:
                     cls.theme[key] = cls.theme[key]["Windows"]
                 else:
                     cls.theme[key] = cls.theme[key]["Linux"]
+
+        # fix name inconsistencies
+        if "CTkCheckbox" in cls.theme.keys():
+            cls.theme["CTkCheckBox"] = cls.theme.pop("CTkCheckbox")
+        if "CTkRadiobutton" in cls.theme.keys():
+            cls.theme["CTkRadioButton"] = cls.theme.pop("CTkRadiobutton")
 
     @classmethod
     def save_theme(cls):